# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto API
  description: |
    The WASAPhoto API interacts with a social image platform. Users can view a stream of photos in reverse chronological order, including upload timestamps, likes, and comments. This stream consists of images from users they follow.

    Users can like and comment on images, with only comment authors able to delete their comments. Blocking users restricts their access to the blocker's information, with the option to unblock at any time.

    Each user has a profile displaying their photos, total uploads, followers, and following. Users can modify usernames, upload/delete photos, and follow/unfollow others.

    Deleting an image removes associated likes and comments.

    Users can search for profiles by username and log in using only their username.
  contact:
    name: Alejandro Ibáñez Pastrana
    email: ibanezpastrana.2116832@studenti.uniroma1.it
    url: https://github.com/aleiis
  version: 1.0.0
servers: [ ]
...
paths:
...
  /users/{user_id}/photos/:
    summary: User's photos
    parameters:
      - $ref: '#/components/parameters/user_id'
    post:
      tags: [ "Photos" ]
      summary: Uploads a photo for the user
      description: |
        The user uploads a photo. The photo is stored in the user's profile.
      operationId: uploadPhoto
      security:
        - bearerAuth: [ ]
      requestBody:
        description: Photo to upload
        content:
          image/png:
            schema:
              description: PNG image
              type: string
              format: binary
              minLength: 1
              maxLength: 1048576
          image/jpeg:
            schema:
              description: JPEG image
              type: string
              format: binary
              minLength: 1
              maxLength: 1048576
        required: true
      responses:
        '201':
          description: Photo uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlobalPhotoId'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/{user_id}/photos/{photo_id}:
    summary: User's photo identified by a unique ID
    parameters:
      - $ref: '#/components/parameters/user_id'
      - $ref: '#/components/parameters/photo_id'
    delete:
      tags: [ "Photos" ]
      summary: Deletes a photo
      description: |
        The user deletes a photo. The photo is removed from the user's profile.
      operationId: deletePhoto
      security:
        - bearerAuth: [ ]
      responses:
        '204':
          description: Photo deleted successfully
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /users/{user_id}/bans/{banned_id}:
    summary: User bans another user
    description: |
      The user identified by user_id bans another user identified by bannedID.
    parameters:
      - $ref: '#/components/parameters/user_id'
      - name: banned_id
        in: path
        description: Identifier of the banned user
        required: true
        schema:
          $ref: '#/components/schemas/UserId'
        example: 0
    delete:
      tags: [ "Banning" ]
      summary: Deletes a ban
      description: |
        Lets the user unban another user.
      operationId: unbanUser
      security:
        - bearerAuth: [ ]
      responses:
        '204':
          description: User unbanned successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags: [ "Banning" ]
      summary: Lets a user check if it has banned another user
      description: |
        If the user has banned the other user the API will return 200 OK as the status code. If not, it will return 404 Not Found.
      operationId: checkBan
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Ban exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ban'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    ErrorMessage:
      title: ErrorMessage
      description: Error message
      type: string
      example: "Internal Server Error"
...
    UserId:
      title: UserId
      description: Unique identifier of a user
      type: integer
      format: int64
      example: 0
...
    UserPhotoId:
      title: UserPhotoId
      description: Identifier of a photo unique per user. Each user has its independent set of photo identifiers.
      type: integer
      format: int64
      example: 0
    GlobalPhotoId:
      title: GlobalPhotoId
      description: Global identifier of a photo. A photo is identified by the user's unique ID and the photo's unique ID per user.
      type: object
      properties:
        owner:
          $ref: '#/components/schemas/UserId'
        id:
          $ref: '#/components/schemas/UserPhotoId'
      example:
        owner: 13
        id: 4
...
    Ban:
      title: Ban
      description: Ban resource
      type: object
      properties:
        ban_issuer:
          $ref: '#/components/schemas/UserId'
        banned_user:
          $ref: '#/components/schemas/UserId'
...
  parameters:
    user_id:
      name: user_id
      in: path
      description: User identifier
      required: true
      schema:
        $ref: '#/components/schemas/UserId'
      example: 0
    photo_id:
      name: photo_id
      in: path
      description: Photo identifier unique per user
      required: true
      schema:
        $ref: '#/components/schemas/UserPhotoId'
      example: 0
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    BadRequestError:
      description: Bad Request
      content:
        text/plain:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
    UnauthorizedError:
      description: Access token is missing or invalid
      content:
        text/plain:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
...
    NotFoundError:
      description: Resource not found
      content:
        text/plain:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
    InternalServerError:
      description: Internal Server Error
      content:
        text/plain:
          schema:
            $ref: '#/components/schemas/ErrorMessage'